Add files via upload
Air quality is measured with the Air Quality Index, or AQI. The AQI works like a thermometer that runs from 0 to 500 degrees. However, instead of showing changes in the temperature, the AQI is a way of showing changes in the amount of pollution in the air. Air quality is a measure of how clean or polluted the air is. Air pollution is one of the most dire environmental problems in the world right now. Constant exposure to it poses incredibly high risks of health issues and premature deaths to the world’s population. According to the World Health Organization (WHO), it estimates that at least 9 out of 10 people across the world are living in places with polluted air. Global air pollution is caused by the burning  of fossil fuels to generate power, as well as emissions from fossil fuels powered motor vehicles. Rich and developed countries contribute a large proportion of global air pollution on the account of their energy generation from non-renewable sources and the amount of vehicles on the road. The Environmental Protection Agency reported that about 68 million tons of air pollution were emitted into the atmosphere in the US in 2020 alone.